#266 Template - Web - SpecFlow - Selenium - System Test

+++ test/Base/AspNetCore/WebProjectServer.cs
@@ -2,10 +2,15 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
+using System.Net.Http;
 using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace Optivem.Framework.Test.AspNetCore
 {
+    // TODO: VC: Rename to ProcessTestServer and return also the client
+
     public class WebProjectServer : IDisposable
     {
         private Process _process;
@@ -14,11 +19,45 @@ namespace Optivem.Framework.Test.AspNetCore
         private const string RunArg = "run";
         private const string ProjectParam = "project";
 
-        public WebProjectServer(string projectPath)
+        private const long MaxRetries = 30;
+
+        private const int Pause = 1000;
+
+        public WebProjectServer(string projectPath, string url, string pingPath, WebPinger pinger)
         {
             if(!File.Exists(projectPath))
             {
                 throw new ArgumentException($"File does not exist: {projectPath}");
+            }
+
+            ProjectPath = projectPath;
+            Url = url;
+            PingPath = pingPath;
+            Pinger = pinger;
+        }
+
+        public string ProjectPath { get; }
+
+        public string Url { get; }
+
+        public string PingPath { get; }
+
+        public WebPinger Pinger { get; }
+
+        public async Task Start()
+        {
+            var running = await Pinger.PingAsync(Url, PingPath);
+
+            // TODO: VC: check
+
+            if(running)
+            {
+                return;
+            }
+
+            if(running)
+            {
+                throw new Exception($"Web server is already running at {Url}");
             }
 
             _process = new Process
@@ -30,20 +69,35 @@ namespace Optivem.Framework.Test.AspNetCore
                     Arguments = $"{RunArg} --project {ProjectPath}",
                 },
             };
-        }
 
-        public string ProjectPath { get; }
-
-        public void Start()
-        {
+            // TODO: VC: Check there is no connection
             _process.Start();
+
+            var currentAttempt = 0;
+
+            while(!running && currentAttempt < MaxRetries)
+            {
+                running = await Pinger.PingAsync(Url, PingPath);
+
+                if(!running)
+                {
+                    currentAttempt += 1;
+                    Thread.Sleep(Pause);
+                }
+            }
+
+            if(!running)
+            {
+                throw new Exception($"Failed to contact web server at {Url}");
+            }
         }
 
         public void Dispose()
         {
-            if(!_process.HasExited)
+            if(_process != null && !_process.HasExited)
             {
                 _process.Kill();
+                _process = null;
             }
         }
     }
